+++ CHANGELOG.md
@@ -92,3 +92,4 @@
 - MVP-458 - Implement signup wizard
 - MVP-450 - update add company logo
 - MVP-512 - Handle user requesting generic content
+- MVP-496 - Handle signup from export plan